fix new clippy lints and try block inference for nightly-2026-04-11
- manual_saturating_arithmetic: use saturating_sub() instead of
  checked_sub().unwrap_or_default()
- unfulfilled_lint_expectations: remove #[expect(clippy::double_parens)]
  no longer needed by construct_uint! macro
- collapsible_match: collapse if-in-match to match guards
- unnecessary_sort_by: use sort_unstable_by_key
- useless_conversion: remove redundant .into_iter() calls
- unnecessary_cast: remove u32-to-u32 casts
- explicit_counter_loop: use zip with range instead of manual counters
- result_large_err: allow on RPC builder closures returning sc_service::Error
- try block type inference: add explicit .map_err() for ReadingError->ProvingError
  conversion (new nightly no longer infers From in try blocks)

+++ crates/pallet-domains/src/block_tree.rs
@@ -1085,7 +1085,7 @@ mod tests {
             // Construct a future receipt
             let mut future_receipt = next_receipt.clone();
             future_receipt.set_domain_block_number(head_receipt_number + 2);
-            future_receipt.set_consensus_block_number(head_receipt_number as u32 + 2);
+            future_receipt.set_consensus_block_number(head_receipt_number + 2);
             ExecutionInbox::<Test>::insert(
                 (
                     domain_id,
@@ -1199,7 +1199,7 @@ mod tests {
             // Construct a future receipt
             let mut future_receipt = next_receipt.clone();
             future_receipt.set_domain_block_number(head_receipt_number + 1);
-            future_receipt.set_consensus_block_number(head_receipt_number as u32 + 1);
+            future_receipt.set_consensus_block_number(head_receipt_number + 1);
 
             ExecutionInbox::<Test>::insert(
                 (
